// ButorRevolutionWebAPI/src/frontend/web/projects/factory/src/app/admin/users/models/users.models.ts
export interface IUserBasicInfoViewModel {
    id?: number;
    name: string;
    roles?: number[];
    status?: string;
    company?: number;
    companyList?: ICompanySelectList;
    address?: IAddressViewModel;
    email?: string;
    phone?: string;
    language?: LanguageTypeEnum;
    addedOn?: string;
    image?: ImageDetailsViewModel;
}
export interface ImageDetailsViewModel {
    containerName?: string;
    fileName?: string;
}
export interface ICompanySelectList {
    id: number;
    name?: string;
}
export interface ILanguageSelectList {
    id: number;
    name?: string;
}

export interface IAddressViewModel {
    address?: string;
    postCode?: number;
    city?: string;
    countryId?: number;
}
// ...
export interface CompanySelectModel {
    id: number;
    name: string;
}
export interface RolesSelectModel {
    id: number;
    name: string;
}
export interface CountryListSelectModel {
    id: number;
    name: string;
}